Measure the radio, and lift the 255-tag wall in firmware
Measured what a tag actually costs the AP by polling one tag's record
at 0.3s and separating two things that get conflated: waiting for a tag
to wake up (5-22s, latency, free) from the AP being busy pushing it an
image (3.9s, identical across all three tags, the number that limits
fleet size). 400 tags serialised is ~26 minutes, which fits a morning
window and only happens on first install anyway -- a normal morning
pushes just the prices that changed.

So the radio was never the constraint. The 255-tag pagination wall is,
and it is fixed here: startPos widens from uint8_t to uint16_t through
tagDBtoJson and the /get_db handler, which could not address a tag past
255 because ?pos=256 truncated to 0 and restarted from the beginning.

Also rewrites /restore_db, which held fsMutex for the entire upload. A
real store's database is a couple of hundred kB; that locked the
filesystem for over a minute and the transfer died before reaching the
branch that releases the mutex, wedging every later restore. It now
takes the mutex per chunk, and answers 500 when loadDB cannot parse the
file instead of 200 from a request handler that never knew what the
upload did.

Both firmware changes compile clean (RAM 24.7%, flash 42.1%) but the S3
is NOT flashed -- the notes and PRICE_API say so at every mention.

Retires the Cloudflare tunnel section in PRICE_API to match reality.

+++ ESP32_AP-Flasher/include/tag_db.h
@@ -108,7 +108,10 @@ extern Config config;
 extern std::vector<tagRecord*> tagDB;
 extern std::unordered_map<int, HwType> hwtype;
 extern std::unordered_map<std::string, varStruct> varDB;
-extern String tagDBtoJson(const uint8_t mac[8] = nullptr, uint8_t startPos = 0);
+// startPos is uint16_t, not uint8_t: /get_db pages roughly 11 tags at a time,
+// so a store with more than 255 tags needs positions a byte cannot hold. It
+// used to truncate, silently restarting pagination from zero.
+extern String tagDBtoJson(const uint8_t mac[8] = nullptr, uint16_t startPos = 0);
 extern bool deleteRecord(const uint8_t mac[8], bool allVersions = true);
 extern void fillNode(JsonObject& tag, const tagRecord* taginfo);
 extern void saveDB(const String& filename);
